+++ code/Lib/variableSpacing/__init__.py
@@ -1,16 +1,15 @@
-import os, glob, shutil
+import os, glob, shutil, json
 from fontParts.world import OpenFont
 
-__version__ = "0.1.2"
-
-KEY = 'com.hipertipo.spacingaxis'
+__version__ = "0.1.3"
 
 '''
 SPACING STATES DATA FORMAT
+--------------------------
 
 1. SPACING LIB
 
-font.lib[f'{KEY}.spacing'] = {
+font.lib[KEY_SPACING] = {
     'default' : {
         'space': {
             'width': 270,
@@ -34,7 +33,7 @@ font.lib[f'{KEY}.spacing'] = {
 
 2. KERNING LIB
 
-font.lib[f'{KEY}.kerning'] = {
+font.lib[KEY_KERNING] = {
     'default' : [
         ('public.kern1.A', 'public.kern2.V', -40),
         ('B', 'J', -20),
@@ -45,33 +44,20 @@ font.lib[f'{KEY}.kerning'] = {
 
 3. COMPONENTS LIB
 
-font.lib[f'{KEY}.components'] = {
+font.lib[KEY_COMPONENTS] = {
     'aacute' : [
-        ('a', (65.0, -10.0)),
-        ('acutecmb', (212.0, 581.0)),
+        ('a', (65.0, -10.0)),          # component
+        ('acutecmb', (212.0, 581.0)),  # component
     ]
     'dollar' : [
-        ('S', (62.0, -10.0)),         # component
-        ((225, 588), 'VyKpf7rL3q'),   # contour
-        ((225, -115), 'ZpTeU9Jwwi'),
+        ('S', (62.0, -10.0)),          # component
+        ((225, 588), 'VyKpf7rL3q'),    # contour
+        ((225, -115), 'ZpTeU9Jwwi'),   # contour
     ],
-}
-
-### OLD FORMAT (DEPRECATED)
-
-font.lib[f'{KEY}.components'] = {
-    'Aacute' : {
-        'A' : (35.0, 0.0),           # component
-        'acutecmb' : (281.0, 726.0), # component
-    }
-    'dollar' : {
-        'S' : (62.0, -10.0),         # component
-        'VyKpf7rL3q' : (225, 588),   # contour
-        'ZpTeU9Jwwi' : (225, -115)   # contour
-    },
 }
 '''
 
+KEY = 'com.hipertipo.spacingaxis'
 KEY_SPACING    = f'{KEY}.spacing'
 KEY_KERNING    = f'{KEY}.kerning'
 KEY_COMPONENTS = f'{KEY}.components'
@@ -80,21 +66,18 @@ KEY_COMPONENTS = f'{KEY}.components'
 # reading
 # -------
 
-def getSpacingStates(font, KEY=KEY):
+def getSpacingStates(font, spacingKey=KEY_SPACING, kerningKey=KEY_KERNING):
     '''
     Get the name of all spacing states available in the font.
 
     ::
         
-        >>> from variableSpacing import getStatesNames
-        >>> spacingStates = getStatesNames(font)
+        >>> from variableSpacing import getSpacingStates
+        >>> spacingStates = getSpacingStates(font)
         >>> print(spacingStates)
         ['default', 'loose', 'tight']
 
     '''
-    spacingKey = f'{KEY}.spacing'
-    kerningKey = f'{KEY}.kerning'
-
     spacingLib = []
     if spacingKey in font.lib:
         if len(font.lib[spacingKey]):
@@ -145,7 +128,7 @@ def getKerningLib(font, kerningKey=KEY_KERNING):
     else:
         return {}
 
-def getComponentsLib(font, spacingKey=KEY_COMPONENTS):
+def getComponentsLib(font, componentsKey=KEY_COMPONENTS):
     '''
     Get the components lib from a given font.
 
@@ -161,8 +144,8 @@ def getComponentsLib(font, spacingKey=KEY_COMPONENTS):
         {'leftMargin': 65, 'width': 524}
 
     '''
-    if spacingKey in font.lib:
-        return font.lib[spacingKey]
+    if componentsKey in font.lib:
+        return font.lib[componentsKey]
     else:
         return {}
 
@@ -287,9 +270,6 @@ def saveKerningToLib(font, spacingState, kerningKey=KEY_KERNING):
 # -------
 # loading
 # -------
-
-# def smartSetMargins(font, lefMargin=None, rightMargin=None, beam=None, componentsKey=KEY_COMPONENTS, verbose=True):
-#     componentsDict = font.lib[componentsKey] if componentsKey in font.lib else {}
 
 def loadSpacingFromLib(font, spacingState, spacingKey=KEY_SPACING, componentsKey=KEY_COMPONENTS, verbose=True):
     '''
@@ -536,11 +516,15 @@ def deleteSpacingStatesLib(font, key=KEY):
         if k.startswith(key):
             del font.lib[k]
 
+deleteSpacingStates = deleteSpacingStatesLib
+
 # ----------
 # generating
 # ----------
 
 def buildSpacingSources(folder, prefix='_', verbose=True):
+
+    ### TO-DO: add support for subfolders
 
     # get all ufo sources in folder
     sources = [ufo for ufo in glob.glob(f'{folder}/*.ufo') if prefix not in os.path.split(ufo)[-1]]
@@ -549,20 +533,21 @@ def buildSpacingSources(folder, prefix='_', verbose=True):
     newSources = []
 
     if verbose:
-        print('building spacing sources...\n')
+        print(f'building spacing sources in {folder}...')
 
     for src in sources:
         srcPath = os.path.join(folder, src)
         srcFont = OpenFont(srcPath, showInterface=False)
 
-        states = getStatesNames(srcFont, KEY)
+        states = getSpacingStates(srcFont)
         if 'default' in states:
             states.remove('default')
 
         if not len(states):
             continue
 
-        ### make sure default state is loaded in primary source
+        ### make sure the default state is loaded in the primary source
+        ### WELL… let’s simply assume it is :)
         # loadSpacingFromLib(srcFont, f'{KEY}.spacing', 'default')
         # loadKerningFromLib(srcFont, f'{KEY}.kerning', 'default')
         # srcFont.save()
@@ -574,7 +559,7 @@ def buildSpacingSources(folder, prefix='_', verbose=True):
                 shutil.rmtree(dstPath)
             # duplicate UFO source
             if verbose:
-                print(f'\tsaving {stateName} as {dstPath}...')
+                print(f"\tsaving spacing state '{stateName}' as {dstPath}...")
             shutil.copytree(srcPath, dstPath)
 
             # set font info in duplicate
@@ -586,7 +571,7 @@ def buildSpacingSources(folder, prefix='_', verbose=True):
             loadKerningFromLib(dstFont, stateName)
 
             # clear spacing states from lib
-            deleteLib(dstFont, KEY) 
+            deleteSpacingStates(dstFont) 
 
             # done!
             dstFont.save()
@@ -600,246 +585,353 @@ def buildSpacingSources(folder, prefix='_', verbose=True):
 
     return newSources
 
-# ----------
-# DEPRECATED
-# ----------
+# -------------
+# export/import
+# -------------
 
-def saveSpacingToLib_old(font, spacingState, spacingKey=KEY_SPACING, componentsKey=KEY_COMPONENTS, verbose=False):
+def exportSpacingStates(font, jsonPath, spacingKey=KEY_SPACING, kerningKey=KEY_KERNING, componentsKey=KEY_COMPONENTS):
     '''
-    Save the font’s current spacing values (width, left margin) to the font lib.
-
-    ::
-
-        from variableSpacing import saveSpacingToLib
-        font = CurrentFont()
-        saveSpacingToLib(font, 'tight')
-
-    '''
-    # ------------
-    # spacing dict
-    # ------------
-
-    # make spacing dict
-
-    D = {}
-    for glyphName in font.glyphOrder:
-        if verbose:
-            print(f'saving spacing data for {glyphName}...')
-
-        glyph = font[glyphName]
-        D[glyphName] = {}
-
-        # store the advance widths of all glyphs
-        D[glyphName]['width'] = glyph.width
-        if verbose:
-            print(f'\twidth: {glyph.width}')
-
-        # store left margin of non-empty glyphs
-        if glyph.leftMargin is not None:
-            D[glyphName]['leftMargin'] = glyph.leftMargin
-            if verbose:
-                print(f'\tleft margin: {glyph.leftMargin}')
-
-        if verbose:
-            print()
-
-    # store spacing dict in lib
-    if spacingKey not in font.lib:
-        font.lib[spacingKey] = {}
-    font.lib[spacingKey][spacingState] = D
-
-    # ---------------
-    # components dict
-    # ---------------
+    Export a font’s spacing states as an external JSON file.
     
-    D = {}
-    for glyphName in font.glyphOrder:
-        glyph = font[glyphName]
-        if not len(glyph.components):
-            continue
-
-        # store offsets to shape origin (bounds)
-        
-        if verbose:
-            print(f'saving component data for {glyphName} ...')
-
-        glyphParts = {}
-
-        for component in glyph.components:
-            cx, cy = component.offset
-            if font[component.baseGlyph].bounds is None:
-                print(f'### empty component for {component.baseGlyph}')
-                continue
-            xMin, yMin, _, _ = font[component.baseGlyph].bounds
-            dx = cx + xMin
-            dy = cy + yMin
-            compID = component.baseGlyph # component.getIdentifier()
-            glyphParts[compID] = dx, dy
-
-        for contour in glyph.contours:
-            xMin, yMin, _, _ = contour.bounds
-            contourID = contour.getIdentifier()
-            glyphParts[contourID] = xMin, yMin
-
-        if verbose:
-            for k, v in glyphParts.items():
-                print(f'\t{k} : {v}')
-
-        D[glyphName] = glyphParts
-
-        if verbose:
-            print()
-
-    # store component dict in lib (same for all states)
-    if componentsKey not in font.lib:
-        font.lib[componentsKey] = {}
-    font.lib[componentsKey] = D
-
-def loadSpacingFromLib_old(font, spacingState, spacingKey=KEY_SPACING, componentsKey=KEY_COMPONENTS, verbose=True):
     '''
-    Load spacing data for a given state from the lib into the font.
+    spacingLib    = getSpacingLib(font, spacingKey=spacingKey)
+    kerningLib    = getKerningLib(font, kerningKey=kerningKey)
+    componentsLib = getComponentsLib(font, componentsKey=componentsKey)
 
+    spacingStates = {
+        spacingKey    : spacingLib,
+        kerningKey    : kerningLib,
+        componentsKey : componentsLib,
+    }
+
+    with open(jsonPath, 'w', encoding='utf-8') as f:
+        json.dump(spacingStates, f, indent=2)
+
+def importSpacingStates(font, jsonPath, spacingKey=KEY_SPACING, kerningKey=KEY_KERNING, componentsKey=KEY_COMPONENTS):
     '''
-    if not spacingKey in font.lib:
-        return
+    Import spacing states from an external JSON file into a font.
+    
+    '''
+    with open(jsonPath, 'r', encoding='utf-8') as f:
+        spacingStates = json.load(f)
 
-    if not spacingState in font.lib[spacingKey]:
-        return
+    spacingLib    = spacingStates.get(spacingKey)
+    kerningLib    = spacingStates.get(kerningKey)
+    componentsLib = spacingStates.get(componentsKey)
 
-    spacingDict    = font.lib[spacingKey][spacingState]
-    componentsDict = font.lib[componentsKey] if componentsKey in font.lib else {}
+    if spacingLib:
+        font.lib[spacingKey] = spacingLib
 
-    # ----------------------------
-    # contours only / empty glyphs
-    # ----------------------------
+    if kerningLib:
+        font.lib[kerningKey] = kerningLib
 
-    for glyphName in font.glyphOrder:
+    if componentsLib:
+        font.lib[componentsKey] = componentsLib
+
+# ------------
+# transforming
+# ------------
+
+from fontPens.marginPen import MarginPen
+from fontParts.world import RGlyph
+from hTools3.modules.pens import DecomposePointPen
+
+def getMargins(glyph, useBeam=False, beamY=400):
+    if useBeam:
+        # decompose glyph
+        g = RGlyph()
+        g.width = glyph.width
+        pen = g.getPointPen()
+        decomposePen = DecomposePointPen(glyph.font, pen)
+        glyph.drawPoints(decomposePen)
+        # get margins from flat copy
+        pen = MarginPen({}, beamY, isHorizontal=True)
+        g.draw(pen)
+        intersections = pen.getAll()
+        if not len(intersections):
+            return
+        leftMargin  = intersections[0]
+        rightMargin = g.width - intersections[-1]
+        return leftMargin, rightMargin
+    else:
+        return glyph.leftMargin, glyph.rightMargin
+
+def smartSetMargins(font, glyphNames, leftMargin=None, rightMargin=None, useBeam=False, beamY=400, verbose=True, setUndo=True):
+
+    # TO-DO: plain margins and margins with beam must be set in one go!
+
+    componentsDict = getComponentsLib(font)
+
+    # --------------------------
+    # contours only / skip empty
+    # --------------------------
+
+    if verbose:
+        print('glyphs with contours only:\n')
+
+    for glyphName in glyphNames:
         glyph = font[glyphName]
+
+        # skip glyphs with components
         if len(glyph.components):
             continue
-        if glyphName not in spacingDict:
+        # skip empty glyphs
+        if not glyph.bounds:
             continue
-        if 'leftMargin' in spacingDict[glyphName]:
-            glyph.leftMargin = spacingDict[glyphName]['leftMargin']
-        glyph.width = spacingDict[glyphName]['width']
+        # skip zero-width glyphs
+        if glyph.width == 0:
+            continue
+
+        if verbose:
+            print(glyphName, end=', ')
+
+        if setUndo:
+            glyph.prepareUndo('smart set margins')
+        oldLeft,  oldRight  = glyph.leftMargin, glyph.rightMargin
+        beamLeft, beamRight = getMargins(glyph, useBeam=useBeam, beamY=beamY)
+
+        # set left margin
+        if leftMargin is not None:
+            leftValue = leftMargin
+            if useBeam:
+                leftValue -= beamLeft - oldLeft
+            glyph.leftMargin = leftValue
+
+        # set right margin
+        if rightMargin is not None:
+            rightValue = rightMargin
+            if useBeam:
+                rightValue -= beamRight - oldRight
+            glyph.rightMargin = rightValue
+
         glyph.changed()
+        if setUndo:
+            glyph.performUndo()
+
+    if verbose:
+        print('\n')
 
     # ---------------------------
     # mixed contours & components
     # ---------------------------
 
-    for glyphName in font.glyphOrder:
+    if verbose:
+        print('glyphs with contours and components:\n')
+
+    for glyphName in glyphNames:
         glyph = font[glyphName]
-        # skip glyphs without components
+
+        # skip glyphs with no components
         if not len(glyph.components):
             continue
+
         # skip empty glyphs
         if not glyph.bounds:
             continue
+
+        # skip zero-width glyphs
+        if glyph.width == 0:
+            continue
+
         # skip glyphs with no contours
         if not len(glyph.contours):
             continue
 
-        glyphParts = componentsDict[glyphName]
+        if verbose:
+            print(glyphName, end=', ')
+
+        if setUndo:
+            glyph.prepareUndo('smart set margins')
+        oldLeft,  oldRight  = glyph.leftMargin, glyph.rightMargin
+        beamLeft, beamRight = getMargins(glyph, useBeam=useBeam, beamY=beamY)
+
+        # set left margin
+        if leftMargin is not None:
+            leftValue = leftMargin
+            if useBeam:
+                leftValue -= beamLeft - oldLeft
+            glyph.leftMargin = leftValue
+
+        # no glyphparts entry for glyph
+        if glyphName not in componentsDict:
+            # set right margin without any adjustment
+            if rightMargin is not None:
+                glyph.rightMargin = rightMargin
+            # move on to the next glyph
+            continue
 
         # adjust component positions
+        glyphParts = componentsDict[glyphName]
+        assert type(glyphParts) is list
+
+        componentsUsed = []
         for comp in glyph.components:
-            compID = comp.baseGlyph
-            if compID not in glyphParts:
-                continue
-            else:
+            # find data for component
+            for i, item in enumerate(glyphParts):
+                # if the item does not start with a string, it’s not a component but a contour
+                if type(item[0]) is not str:
+                    continue
+
+                # match item to component
+                if comp.baseGlyph != item[0]:
+                    continue
+
+                # it’s possible to have multiple components of the same base glyph
+                # each instance has its own entry; this makes sure that we only use each item once
+                if i in componentsUsed:
+                    continue
+
                 cx, cy = comp.offset
+                if comp.baseGlyph not in font:
+                    continue
+
                 baseGlyph = font[comp.baseGlyph]
                 xMin, yMin, _, _ = baseGlyph.bounds
                 dx = xMin + cx
                 dy = yMin + cy
-
-                sx, sy = glyphParts[compID]
+                sx, sy = item[1]
                 diffx = sx - dx
                 diffy = sy - dy
-
+                componentsUsed.append(i)
+                if diffx == 0 and diffy == 0:
+                    continue
                 comp.moveBy((diffx, diffy))
+                break
 
         # adjust contour positions
         for contour in glyph.contours:
             contourID = contour.getIdentifier()
-            if contourID not in glyphParts:
-                continue
-            else:
+            for i, item in enumerate(glyphParts):
+                # if the item does not start with a tuple, it’s not a contour (but a component)
+                if type(item[0]) is not tuple:
+                    continue
+
+                # match item to contour
+                if contourID != item[1]:
+                    continue
+
                 dx, dy, _, _ = contour.bounds
-                sx, sy = glyphParts[contourID]
+                sx, sy = item[0]
                 diffx = sx - dx
                 diffy = sy - dy
                 contour.moveBy((diffx, diffy))
-        
+
         glyph.changed()
 
-        leftMargin = spacingDict[glyph.name].get('leftMargin')
+        # set left margin
         if leftMargin is not None:
-            glyph.leftMargin = leftMargin
-            glyph.changed()
+            glyph.leftMargin = leftValue
 
-        glyphWidth = spacingDict[glyph.name].get('width')
-        glyph.width = glyphWidth
+        # set right margin
+        if rightMargin is not None:
+            rightValue = rightMargin
+            if useBeam:
+                rightValue -= beamRight - oldRight
+            glyph.rightMargin = rightValue
+
         glyph.changed()
+        if setUndo:
+            glyph.performUndo()
 
-    # ---------------
+    if verbose:
+        print('\n')
+
+    # ----------------
     # components only
-    # ---------------
+    # ----------------
+    if verbose:
+        print('glyphs with components only:\n')
 
-    for glyphName in font.glyphOrder:
+    for glyphName in glyphNames:
         glyph = font[glyphName]
+
         # skip glyphs without components
         if not len(glyph.components):
             continue
+
         # skip empty glyphs
         if not glyph.bounds:
             continue
+
+        # skip zero-width glyphs
+        if glyph.width == 0:
+            continue
+
         # skip glyphs with contours
         if len(glyph.contours):
             continue
-            
-        glyphParts = componentsDict[glyphName] if glyphName in componentsDict else {}
+
+        if verbose:
+            print(glyphName, end=', ')
+        if setUndo:
+            glyph.prepareUndo('smart set margins')
+        oldLeft,  oldRight  = glyph.leftMargin, glyph.rightMargin
+        beamLeft, beamRight = getMargins(glyph, useBeam=useBeam, beamY=beamY)
+
+        glyphParts = componentsDict[glyphName] if glyphName in componentsDict else []
+        assert type(glyphParts) is list
 
         glyph.leftMargin = 0
         glyph.changed()
 
+        componentsUsed = []
         for comp in glyph.components:
-            compID = comp.baseGlyph # comp.getIdentifier()
-            if compID not in glyphParts:
-                continue
-            else:
+            # find data for component
+            for i, item in enumerate(glyphParts):
+
+                # if the item does not start with a string, it’s not a component (but a contour)
+                if type(item[0]) is not str:
+                    continue
+
+                # match item to component
+                if comp.baseGlyph != item[0]:
+                    continue
+
+                # it’s possible to have multiple components of the same base glyph
+                # each instance has its own entry
+                # this makes sure that we only use each item once
+                if i in componentsUsed:
+                    continue
                 cx, cy = comp.offset
                 if comp.baseGlyph not in font:
                     continue
+
                 baseGlyph = font[comp.baseGlyph]
                 xMin, yMin, _, _ = baseGlyph.bounds
+
                 # current position
                 dx = xMin + cx
                 dy = yMin + cy
+
                 # saved position
-                sx, sy = glyphParts[compID]
-                
+                sx, sy = item[1]
                 diffx = sx - dx
                 diffy = sy - dy
-
+                componentsUsed.append(i)
                 if diffx == 0 and diffy == 0:
                     continue
 
+                # update component position
                 comp.moveBy((diffx, diffy))
                 glyph.changed()
-            
-        if glyph.name not in spacingDict:
-            continue
+                break
 
-        leftMargin = spacingDict[glyph.name].get('leftMargin')
         if leftMargin is not None:
-            glyph.leftMargin = leftMargin
+            leftValue = leftMargin
+            if useBeam:
+                leftValue -= beamLeft - oldLeft
+            glyph.leftMargin = leftValue
             glyph.changed()
 
-        glyphWidth = spacingDict[glyph.name].get('width')
-        glyph.width = glyphWidth
-        glyph.changed()
+        if rightMargin:
+            rightValue = rightMargin
+            if useBeam:
+                rightValue -= beamRight - oldRight
+            glyph.rightMargin = rightValue
+            glyph.changed()
 
-getStatesNames = getSpacingStates
-deleteLib = deleteSpacingStatesLib
+        if setUndo:
+            glyph.performUndo()
+
+    # done
+    font.changed()