+++ docs/_topics/changelog.md
@@ -37,7 +37,8 @@ types of changes:
 0.1.3
 -----
 
-- <span class='badge'>Added</span> Adding a new tool [Smart Margins] to change glyph margins without breaking component positioning.
+- <span class='badge'>Added</span> Adding functions to export/import spacing states to/from external JSON files in the SpacingStates tool.
+- <span class='badge'>Added</span> Adding a new [Smart Margins] tool to change glyph margins without breaking component positioning.
 - <span class='badge'>Changed</span> Moving code to save/load components lib to standalone functions.
 
 [Smart Margins]: ../reference/smart-margins-tool/